# backend/src/openapi.yaml
openapi: 3.0.3
info:
  title: Leaflet URL Shortener API
  description: Privacy-first URL shortener backend API
  version: 1.0.0
servers:
  - url: http://localhost:3001
    description: Configured API server
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: connect.sid
    BearerAuth:
      type: http
      scheme: bearer
      description: OAuth 2.0 access token
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
    OAuthError:
      type: object
      properties:
        error:
          type: string
        error_description:
          type: string
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Seconds until the access token expires
        refresh_token:
          type: string
        scope:
          type: string
    OAuthClient:
      type: object
      properties:
        clientId:
          type: string
        clientSecret:
          type: string
          nullable: true
          description: Raw secret shown only at registration time; null for public clients
        name:
          type: string
        isPublic:
          type: boolean
        redirectUris:
          type: array
          items:
            type: string
        scopes:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    ConsentedApp:
      type: object
      properties:
        clientId:
          type: string
        name:
          type: string
        scopes:
          type: array
          items:
            type: string
        grantedAt:
          type: string
          format: date-time
paths:
...
  /oauth/token:
    post:
      summary: OAuth 2.0 Token endpoint
      description: |
        Exchanges an authorization code for tokens, or refreshes tokens.
        Uses application/x-www-form-urlencoded request body.
        No CSRF token required.
      tags: [OAuth]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [grant_type, client_id]
              properties:
                grant_type:
                  type: string
                  enum: [authorization_code, refresh_token]
                client_id:
                  type: string
                client_secret:
                  type: string
                  description: Required for confidential clients
                code:
                  type: string
                  description: Required for authorization_code grant
                redirect_uri:
                  type: string
                  description: Required for authorization_code grant; must match original
                code_verifier:
                  type: string
                  description: PKCE verifier; required when code_challenge was used
                refresh_token:
                  type: string
                  description: Required for refresh_token grant
      responses:
        '200':
          description: Token response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid request or grant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '401':
          description: Invalid client credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
  /oauth/revoke:
    post:
      summary: OAuth 2.0 Token Revocation endpoint (RFC 7009)
      description: |
        Revokes an access or refresh token. Always returns 200.
        Uses application/x-www-form-urlencoded request body.
      tags: [OAuth]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token, client_id]
              properties:
                token:
                  type: string
                token_type_hint:
                  type: string
                  enum: [access_token, refresh_token]
                client_id:
                  type: string
                client_secret:
                  type: string
      responses:
        '200':
          description: Revocation acknowledged (regardless of whether the token existed)
  /oauth/apps:
    get:
      summary: List apps the authenticated user has authorized
      tags: [OAuth]
      security:
        - sessionCookie: []
        - BearerAuth: []
      responses:
        '200':
          description: List of consented applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConsentedApp'
        '401':
          description: Authentication required
    post:
      summary: Register a new OAuth client application
      tags: [OAuth]
      security:
        - sessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, redirectUris, scopes, isPublic]
              properties:
                name:
                  type: string
                redirectUris:
                  type: array
                  items:
                    type: string
                scopes:
                  type: array
                  items:
                    type: string
                    enum: [shorten:create, shorten:create:never, shorten:create:alias,
                           urls:read, urls:delete, users:read, users:write,
                           user:read, oauth:apps:read, oauth:apps:write, admin:*]
                isPublic:
                  type: boolean
      responses:
        '201':
          description: Client registered; clientSecret is shown only once
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthClient'
        '400':
          description: Validation error
        '401':
          description: Authentication required
  /oauth/apps/{clientId}:
    delete:
      summary: Revoke an OAuth client and all its active tokens
      description: Owner or admin only.
      tags: [OAuth]
      security:
        - sessionCookie: []
        - BearerAuth: []
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Application revoked
        '401':
          description: Authentication required
        '403':
          description: Forbidden
        '404':
          description: Client not found
  /api/shorten:
    post:
      summary: Create a short URL
      tags: [URLs]
      security:
        - sessionCookie: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, ttl]
              properties:
                url:
                  type: string
                  format: uri
                  description: The URL to shorten
                ttl:
                  type: string
                  enum: [5m, 1h, 24h, never]
                  description: Time-to-live. "never" is admin only.
                alias:
                  type: string
                  minLength: 3
                  maxLength: 50
                  description: Custom short code alias. Privileged/admin only.
      responses:
        '201':
          description: Short URL created
          content:
            application/json:
              schema:
                type: object
                properties:
                  shortCode:
                    type: string
                  shortUrl:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
                    nullable: true
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Alias already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded (anonymous users)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /s/{code}:
    get:
      summary: Canonical short-link redirect
      tags: [URLs]
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Temporary redirect to the original URL
        '404':
          description: Short URL not found or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...